fix and improve SentMessage.delete()
If there is a problem in delete() method, the sent message will not be save.
We save it first because it's a simpler operation and we don't want to lose
the sent message's ts.
Then we delet the other messages. from the channel.
Also there is a new force_delete method, which tries to delete the slack
message, but will delete it from the database even if it fails.