Implement new crontab form
with as user-friendly messages as possible.
Don't bother the user with irrelevant messages, but show as much as needed.
For example: if they specify both fields, only show that error message and
don't bother with either of the fields is starting with the wrong character.

Added django-widget-tweaks for easier rendering of form fields.

Messages needs to be posted with the as_user flag
to be able to delete them later.

Form validation need to be not the same on the admin page than this user facing
page, because otherwise it would be not possible to save an existing model.