Fully functional Dockerfile
Code is inside /app, working directory is /app/data.
Data is only stored there so it can be bind-mounted to the host to keep the
database. The memory mapped file could not be bind mounted, because it would
crash uwsgi. It is put under /tmp, so it can still be re-used if the container
is not deleted on stop.

Generate a secret key if the SECRET_KEY environment variable is not set
or there is no secret_key.txt file. If there is one, read from that.

Apply migrations before every start.