Save review requests from messages, new model
The ReviewRequest model is to store requests other than scheduled gerrit
queries, usually sent to a team's channel. They will be tied to the first
Crontab which has the same channel id where they has been sent. This can be
changed later.

The gerrit_query field of the Crontab can be empty, for channels which only want
to store Review Requests, but no scheduled gerrit queries should be sent.
They will not be loaded for the bot.