Separated RTM Api handler to a new class
It is totally different than any other part of the API, it's never synchronous
and only can do things when already connected, has special methods and the order
they are called is important.

Also make the message handling explicit without giving a callback. This is more
flexible because messages can be sent any time, not just during handling an
event. It's also simpler, because the user is responsible for the message
handling, there is no need for functiools.partial mambo-jambo anymore.

channel parameters has been renamed to channel_id, those are more approriate.

The "async with await" construct is a bit unfortunate, but makes sense,
because the RTM api can be connected concurrently.