More sophisticated reply to a review request
Now it is able to tell the reason why it is rejected.

Parsing the queries from the urls later makes the code simpler, because it is
only needed to pass around the found Gerrit urls, not a tuple.

New API method: get_permalink, which is needed for referencing the previous
review request message by url.