bot name appear differently in "message" type
than in "desktop_notification" type messages. The bot id is used instead of the
full name, so we need to save the initial response to rtm.connect and use that
to check if the bot has been mentioned.

We can also use it for checkind own user id instead getting it from database.