Run bot.py in an event loop with loop.call_at
This doesn't always work with Python 3.6.6, because the when argument cannot be
arbitrary big. This will be reverted in the next commit, only left here to show
up in the Git history. The problem is fixed in Python 3.7.1, but Alpine Linux
latest Python is 3.6.6.

This Exception is raised:
Traceback (most recent call last):
  File "/app/bot.py", line 250, in <module>
    main()
  File "/app/bot.py", line 246, in main
    loop.run_forever()
  File "/usr/lib/python3.6/asyncio/base_events.py", line 422, in run_forever
    self._run_once()
  File "/usr/lib/python3.6/asyncio/base_events.py", line 1379, in _run_once
    event_list = self._selector.select(timeout)
  File "/usr/lib/python3.6/selectors.py", line 445, in select
    fd_event_list = self._epoll.poll(timeout, max_ev)
OverflowError: timeout is too large