Wait for message deletion in a channel
loop.create_task would be too fast, the Slack client vibrates a lot and there
would be two messages at the same time for a short moment, which is way more
disturbing, so we wait for the deletion to finish in a channel instead.

Multiple messages can be posted at the same time to multiple channels, but
that's not a problem, because a user can only see one channel at a time.