Allow only ReviewRequests for channels
It is possible to use Slackbot in a way, where no cron entry is configured,
only the but is invited to a channel and parse and save sent ReviewRequests,
without downloading a fixed set of changes hourly.

An empty gerrit_query field means the Crontab object is only used to store
the channel_id.

Another solution could be a separate Channel abstraction and ReviewRequests
and Crontabs would reference that, but I don't think it's necessary at this
point yet.